app/requirements.txt
shiny>=0.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
Pillow>=10.0.0

